# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task API",
        "description": "Task API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/tasks": {
            "get": {
                "summary": "Get list of tasks",
                "operationId": "be68905d866e856db6c738172b8d929a",
                "responses": {
                    "200": {
                        "description": "Task list"
                    }
                }
            },
            "post": {
                "summary": "Create a new task",
                "operationId": "ebbaabc58e3db61374a57c53544959b2",
                "parameters": [
                    {
                        "name": "X-Requested-With",
                        "in": "header",
                        "description": "Ajax request",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "XMLHttpRequest"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Задача1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Задача1 описание"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-01-20T15:00:00"
                                    },
                                    "create_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-01-20T15:00:00"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "enum": [
                                            "низкий",
                                            "средний",
                                            "высокий"
                                        ],
                                        "example": "высокий"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Работа"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "выполнена",
                                            "не выполнена"
                                        ],
                                        "example": "не выполнена"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "summary": "Get task by id",
                "operationId": "32c40e2d1129b80c1dcabd538df63a33",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the task",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task found"
                    },
                    "404": {
                        "description": "Task not found"
                    }
                }
            },
            "put": {
                "summary": "Update task by id",
                "operationId": "c36c959b634c99e1e02ad2f3f2c36a2c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the task",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-Requested-With",
                        "in": "header",
                        "description": "Ajax request",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "XMLHttpRequest"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Задача2"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Задача2 описание обновленное"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-01-25T18:00:00"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "enum": [
                                            "низкий",
                                            "средний",
                                            "высокий"
                                        ],
                                        "example": "высокий"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "выполнена",
                                            "не выполнена"
                                        ],
                                        "example": "не выполнена"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task updated"
                    },
                    "404": {
                        "description": "Task not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "summary": "Delete task by id",
                "operationId": "d81cc7975ea62176b4ef8f74ed95168d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the task",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task deleted"
                    },
                    "404": {
                        "description": "Task not found"
                    }
                }
            }
        }
    }
}